Add major version upgrade agreement mechanism
Interim checkin for a mechanism that gets user approval when there is a
major version change (e.g. 18.02.xx -> 18.08.xx).  When detected, the
user hears more details about the impact and is asked for confirmation
before the upgrade is allowed to continue.  This also updates the
USER/mycroft.conf with the "max_allowed_core_version" after the agreement,
which is used in the update cron job thereafter.

Also cleaned up some of the function calls and improved pronunciation of
the version numbers.